+++ tests/src/chanjo2/endpoints/test_report.py
@@ -6,7 +6,7 @@ from fastapi.testclient import TestClient
 from requests.models import Response
 
 from chanjo2.constants import GENE_LISTS_NOT_SUPPORTED_MSG
-from chanjo2.demo import DEMO_COVERAGE_QUERY_DATA
+from chanjo2.demo import DEMO_COVERAGE_QUERY_DATA, DEMO_COVERAGE_QUERY_FORM
 
 
 def test_demo_report(client: TestClient, endpoints: Type):
@@ -38,11 +38,11 @@ def test_report_request_no_genes(client: TestClient, endpoints: Type):
     # THEN the response should return the expected error code and message
     assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
     result = response.json()
-    assert GENE_LISTS_NOT_SUPPORTED_MSG in result["detail"][0]["msg"]
+    assert GENE_LISTS_NOT_SUPPORTED_MSG in result["detail"]
 
 
 def test_report_json_data(client: TestClient, endpoints: Type):
-    """Test the coverage report endpoint by providing json data in POST request."""
+    """Test the coverage report endpoint by providing json data in a POST request."""
 
     # GIVEN a query with json data to the report endpoint
     response: Response = client.post(
@@ -51,6 +51,22 @@ def test_report_json_data(client: TestClient, endpoints: Type):
     )
 
     # THEN the request should be successful
+    assert response.status_code == status.HTTP_200_OK
+
+    # And return an HTML page
+    assert response.template.name == "report.html"
+
+
+def test_report_form_data(client: TestClient, endpoints: Type):
+    """Test the coverage report endpoint by providing application/x-www-form-urlencoded data in a POST request."""
+
+    # GIVEN a query with application/x-www-form-urlencoded data to the report endpoint
+    response: Response = client.post(
+        endpoints.REPORT,
+        data=DEMO_COVERAGE_QUERY_FORM,
+        headers={"Content-Type": "application/x-www-form-urlencoded"},
+    )
+    # Then the request should be successful
     assert response.status_code == status.HTTP_200_OK
 
     # And return an HTML page