+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [1.6]
+### Added
+- Coverage report and genes coverage overview endpoints now accept also requests with application/x-www-form-urlencoded data
 ### Fixed
 - Faster genes overview report loading
 - Broken GitHub action due to d4tools failing to install using cargo